Remove OsSignal trait entirely.
None of this code was ever used on Windows, and thus contributes to
masking what's truly common across the platforms.
